Large update.
Networking implemented. The player character was redesigned to fit networking standards. The player lobby was implemented and players can now connect via local host or through the internets. most of the other scripts are unusable or have not been added/modified. This needs to be redone before we can make our next steps.